Add IN_CLUSTER setting to Kubernetes driver
The kubernetes driver defaults KUBECONFIG to ~/.kube/config when not
set. This prevents the incluster configuration from being used.

To preserve the existing behavior, I have added IN_CLUSTER which
indicates that ambient kubernetes configuration should be used. This is
intended for when the cnab process is running inside a pod.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>