Log error when k8s driver can't stream logs
Currently we are silently ignoring errors when we cannot stream the pod logs. Which makes it super hard to troubleshoot why it didn't work, usually RBAC because you don't have get+list for pods/log. This prints a message in the logs indicating there was a problem and allows the job to finish.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>